Rename RefreshIpCommand.py to refresh_ip.py
connect #47
