+++ vCenterShell/commands/refresh_ip.py
